Improve documentation
primarily adds safety notes for unsafe file-backed maps. Such a general
note doesn't have to be "authoritative" or a treatise on the nuances of
every supported platform in order to give users a useful starting point
in assessing the (un-)safety. GitHub cc: #70

Also switches docs to use the new rustdoc link format (RFC 1946), and
cleans up some of the copy.